mikkel to states and progress bar to training (nuked graphics.py)
